+++ wulu/src/emoticons.js
@@ -53,7 +53,9 @@ function Emoticons(_emotes=emotes) {
   }
 
   CommandParser.parse = function(message, room, user, connection, levelsDeep) {
-    if (message.substr(0, 3) === '/pm') return CommandParser.originalParse(message, room, user, connection, levelsDeep);
+    if ((message.charAt(0) === '/' && message.charAt(1) !== '/') || message.charAt(0) === '!') {
+      return CommandParser.originalParse(message, room, user, connection, levelsDeep);
+    }
 
     let match = false;
     let len = emotes_keys.length;