+++ wulu/src/usergroups.js
@@ -2,8 +2,7 @@ import mongoose from 'mongoose';
 
 let userSchema = new mongoose.Schema({
   name: { type: String, lowercase: true, unique: true },
-  money: { type: Number, default: 0 },
-  symbol: String
+  group: { type: String, default: '' }
 });
 
-export default mongoose.model('user', userSchema);
+export default mongoose.model('usergroup', userSchema);