+++ wulu/src/commands/wulu.js
@@ -1,6 +1,5 @@
 import jf from 'jsonfile';
 import path from 'path';
-import util from 'util';
 
 export default wulu;