+++ wulu/src/commands/regdate.js
@@ -15,7 +15,7 @@ function regdate() {
       let options = {
         host: 'www.pokemonshowdown.com',
         port: 80,
-        path: '/forum/~' + target
+        path: '/users/~' + target
       };
 
       let content = '';