+++ Gruntfile.js
@@ -1,5 +1,3 @@
-// jscs:disable
-
 var fs = require('fs');
 
 module.exports = function(grunt) {