+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Seperate users Mongo collections into users and usergroups
     - usernames are now userids in the users collections
   * Add stafflist command
+  * dev-deps: eslint@^0.19.0
 
 1.9.3 / 2015-04-09
 ==================